+++ com.unity.render-pipelines.lightweight/CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added camera additional data component to control shadows, depth and color texture.
+- pipeline now uses XRSEttings.eyeTextureResolutionScale as renderScale when in XR.
+
+### Changed
+- Shadow rendering has been optimized for the Mali Utgard architecture by removing indexing and avoiding divisions for orthographic projections. This reduces the frame time by 25% on the Overdraw benchmark.
+- Removed 7x7 tent filtering when using cascades.
+- Updated lightweight asset inspector to show capabilities and general settings groups. Removed custom shader stripping.
+- Screenspace shadow resolve is now only done when rendering shadow cascades.
 
 ## [2.0.3-preview]